Add files via upload
My very first project to get started on understanding how python can interact with a system by running a script as a background process. The goal is to kill tasks that fill up memory, disc or network. The project as I said, is just to learn, as this can be done very easily with basic windows tools like command line or registry.
Until now I still cant run the script, as I think this is beyond my skills or knowledge, I would appreciate if anyone could tell me whats wrong with it or if there is a much simpler way to make this happen.